requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-Hashing==1.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
Werkzeug==3.0.1
WTForms==3.1.2
Flask-WTF==1.2.1
packaging==24.0
plotly==5.21.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tenacity==8.2.3